# Implement optional ring buffer for container logs