--- a/daemon/logger/copier.go
+++ b/daemon/logger/copier.go
@@ -76,15 +76,14 @@ func (c *Copier) copySrc(name string, src io.Reader) {
 			}
 			// Break up the data that we've buffered up into lines, and log each in turn.
 			p := 0
-			for q := bytes.Index(buf[p:n], []byte{'\n'}); q >= 0; q = bytes.Index(buf[p:n], []byte{'\n'}) {
+			for q := bytes.IndexByte(buf[p:n], '\n'); q >= 0; q = bytes.IndexByte(buf[p:n], '\n') {
 				select {
 				case <-c.closed:
 					return
 				default:
-					msg := &Message{
-						Source:    name,
-						Timestamp: time.Now().UTC(),
-					}
+					msg := NewMessage()
+					msg.Source = name
+					msg.Timestamp = time.Now().UTC()
 					msg.Line = append(msg.Line, buf[p:p+q]...)
 
 					if logErr := c.dst.Log(msg); logErr != nil {
@@ -98,11 +97,9 @@ func (c *Copier) copySrc(name string, src io.Reader) {
 			// noting that it's a partial log line.
 			if eof || (p == 0 && n == len(buf)) {
 				if p < n {
-					msg := &Message{
-						Source:    name,
-						Timestamp: time.Now().UTC(),
-						Partial:   true,
-					}
+					msg := NewMessage()
+					msg.Source = name
+					msg.Timestamp = time.Now().UTC()
 					msg.Line = append(msg.Line, buf[p:n]...)
 					msg.Partial = true