--- a/daemon/logger/copier.go
+++ b/daemon/logger/copier.go
@@ -47,7 +47,6 @@ func (c *Copier) copySrc(name string, src io.Reader) {
 	buf := make([]byte, bufSize)
 	n := 0
 	eof := false
-	msg := &Message{Source: name}
 
 	for {
 		select {
@@ -78,13 +77,16 @@ func (c *Copier) copySrc(name string, src io.Reader) {
 			// Break up the data that we've buffered up into lines, and log each in turn.
 			p := 0
 			for q := bytes.Index(buf[p:n], []byte{'\n'}); q >= 0; q = bytes.Index(buf[p:n], []byte{'\n'}) {
-				msg.Line = buf[p : p+q]
-				msg.Timestamp = time.Now().UTC()
-				msg.Partial = false
 				select {
 				case <-c.closed:
 					return
 				default:
+					msg := &Message{
+						Source:    name,
+						Timestamp: time.Now().UTC(),
+					}
+					msg.Line = append(msg.Line, buf[p:p+q]...)
+
 					if logErr := c.dst.Log(msg); logErr != nil {
 						logrus.Errorf("Failed to log msg %q for logger %s: %s", msg.Line, c.dst.Name(), logErr)
 					}
@@ -96,9 +98,14 @@ func (c *Copier) copySrc(name string, src io.Reader) {
 			// noting that it's a partial log line.
 			if eof || (p == 0 && n == len(buf)) {
 				if p < n {
-					msg.Line = buf[p:n]
-					msg.Timestamp = time.Now().UTC()
+					msg := &Message{
+						Source:    name,
+						Timestamp: time.Now().UTC(),
+						Partial:   true,
+					}
+					msg.Line = append(msg.Line, buf[p:n]...)
 					msg.Partial = true
+
 					if logErr := c.dst.Log(msg); logErr != nil {
 						logrus.Errorf("Failed to log msg %q for logger %s: %s", msg.Line, c.dst.Name(), logErr)
 					}