… daemon/logger/logger_test.go
… daemon/logger/splunk/splunk_test.go