@LK4D4 2016-11-23
@cpuguy83 maybe embed it to log-opt?