@cpuguy83 2016-11-23
> @cpuguy83 maybe embed it to log-opt?

@LK4D4 The reason I didn't want to do that is `log-opt` tends to be options for the driver, where as this is changing how docker itself handles logs.