@cpuguy83 2016-11-26
@justincormack Implementing this in `github.com/pkg/iotutils.BytesPipe` is fairly trivial I think, but wondering what's better, drop random byte slices or full messages.