@cpuguy83 2016-11-26
Although certainly implementing in BytesPipe is significantly faster since we don't need to do any extra allocations, where as each call to `Log()` as implemented here requires 1 allocation.

I went ahead and added an implementation in `BytesPipe` just to look at, presented without tests or benchmarks at the moment.