@justincormack 2016-11-26
Best to drop full messages I think. They might be JSON or something and
truncating would be unhelpful.

On 26 Nov 2016 2:35 p.m., "Brian Goff" <notifications@github.com> wrote:

> @justincormack <https://github.com/justincormack> Implementing this in
> github.com/pkg/iotutils.BytesPipe is fairly trivial I think, but
> wondering what's better, drop random byte slices or full messages.
>
> —
> You are receiving this because you were mentioned.
> Reply to this email directly, view it on GitHub
> <https://github.com/docker/docker/pull/28762#issuecomment-263066514>, or mute
> the thread
> <https://github.com/notifications/unsubscribe-auth/AAdcPPpe7Uf5tLMOCojYaCpkvPg0em68ks5rCEOugaJpZM4K6tX6>
> .
>