@cpuguy83 2016-11-28
Removed the commit with the bytespipe changes.