@LK4D4 2016-11-28
Cond based buffer is kinda cool. Will review later today.