@cpuguy83 2016-11-28
Benchmarks:

```
BenchmarkRingLoggerThroughputNoReceiver-8           	300000000	       135 ns/op	 229.05 MB/s	     133 B/op	       1 allocs/op
BenchmarkRingLoggerThroughputWithReceiverDelay0-8   	200000000	       212 ns/op	 146.06 MB/s	     113 B/op	       1 allocs/op
BenchmarkRingLoggerThroughputConsumeDelay1-8        	300000000	       130 ns/op	 237.55 MB/s	     133 B/op	       1 allocs/op
BenchmarkRingLoggerThroughputConsumeDelay10-8       	300000000	       132 ns/op	 234.68 MB/s	     133 B/op	       1 allocs/op
BenchmarkRingLoggerThroughputConsumeDelay50-8       	300000000	       126 ns/op	 244.84 MB/s	     133 B/op	       1 allocs/op
BenchmarkRingLoggerThroughputConsumeDelay100-8      	300000000	       131 ns/op	 235.54 MB/s	     133 B/op	       1 allocs/op
BenchmarkRingLoggerThroughputConsumeDelay300-8      	300000000	       130 ns/op	 237.30 MB/s	     133 B/op	       1 allocs/op
BenchmarkRingLoggerThroughputConsumeDelay500-8      	300000000	       128 ns/op	 240.39 MB/s	     133 B/op	       1 allocs/op
```

The happiest case (no receiver delay) is significantly slower. I think this is due to lock contention, but probably does not matter much unless the log driver is doing it's own queueing.
Even with a 1ms delay, throughput is back up.