@cpuguy83 2016-11-29
Updated benchmarks after fixing copy logic.
The no-delay scenario is really slow here.
Specifically, when I use a sync.Pool to handling messages when copying/done with messages, it drops from around 125MB/sec to 75MB/sec, but the other cases go from ~170MB/sec to 220MB/sec

```
BenchmarkRingLoggerThroughputNoReceiver-8           	300000000	       146 ns/op	 211.68 MB/s	      69 B/op	       1 allocs/op
BenchmarkRingLoggerThroughputWithReceiverDelay0-8   	100000000	       388 ns/op	  79.84 MB/s	      52 B/op	       1 allocs/op
BenchmarkRingLoggerThroughputConsumeDelay1-8        	300000000	       170 ns/op	 181.94 MB/s	      69 B/op	       1 allocs/op
BenchmarkRingLoggerThroughputConsumeDelay10-8       	300000000	       153 ns/op	 201.78 MB/s	      69 B/op	       1 allocs/op
BenchmarkRingLoggerThroughputConsumeDelay50-8       	300000000	       138 ns/op	 224.37 MB/s	      69 B/op	       1 allocs/op
BenchmarkRingLoggerThroughputConsumeDelay100-8      	300000000	       146 ns/op	 211.85 MB/s	      69 B/op	       1 allocs/op
BenchmarkRingLoggerThroughputConsumeDelay300-8      	300000000	       143 ns/op	 216.44 MB/s	      69 B/op	       1 allocs/op
BenchmarkRingLoggerThroughputConsumeDelay500-8      	300000000	       144 ns/op	 214.35 MB/s	      69 B/op	       1 allocs/op
```