@cpuguy83 2016-12-09
Ok, what I've done is gone ahead and make the copier handle the allocations.
This simplifies some things for log drivers that are buffering (and I'm not confident that prior to this PR we are doing the right thing w.r.t. making copies of the log message in buffered drivers).

It evens out the performance of the ring buffer but causes more allocations and more bytes being moved per op.
Still the performance here is a bit better than before #22982.
For comparison:

### Before #22982 (<= 1.12)
| Size | Ops | Speed | Memory | Allocs |
| --- | --- | --- | --- | --- |
| 1K | 10000 | 141169 ns/op | 16351 B/op | 39 allocs/op |
| 2K | 20000 | 98573 ns/op | 26577 B/op | 39 allocs/op |
| 4K | 10000 | 105285 ns/op | 47058 B/op | 39 allocs/op |
| 8K | 10000 | 138987 ns/op | 129302 B/op | 59 allocs/op |
| 16K | 10000 | 171856 ns/op | 294591 B/op | 100 allocs/op |
| 32K | 5000 | 307481 ns/op | 624206 B/op | 152 allocs/op |
| 64K | 2000 | 582192 ns/op | 1283438 B/op | 245 allocs/op |
| 128K | 1000 | 1212861 ns/op | 2601906 B/op | 422 allocs/op |
| 256K | 1000 | 2267578 ns/op | 5238848 B/op | 766 allocs/op |

### After #22982 (1.13.0-rcX)
| Size | Ops | Speed | Memory |Allocs |
| --- | --- | --- |  --- |  --- |
| 1K | 10000 | 101876 ns/op |  17664 B/op |  20 allocs/op | 
| 2K | 10000 | 100582 ns/op |  17664 B/op |  20 allocs/op | 
| 4K | 10000 | 103867 ns/op |  17664 B/op |  20 allocs/op | 
| 8K | 10000 | 117245 ns/op |  17664 B/op |  20 allocs/op | 
| 16K | 10000 | 120569 ns/op |  17664 B/op | 20 allocs/op | 
| 32K | 5000 | 236115 ns/op | 17664 B/op |  20 allocs/op | 
| 64K | 3000 | 434702 ns/op |  17664 B/op |  20 allocs/op |
| 128K | 2000 | 839920 ns/op |  17664 B/op |  20 allocs/op | 
| 256K | 1000 | 1611549 ns/op |  17664 B/op |  20 allocs/op |

### This PR
| Size | Ops | Speed | Memory |Allocs |
| --- | --- | --- | --- | --- |
| 1K                  |                     300000         |    40589 ns/op      |     28800 B/op   |      39 allocs/op |
| 2K               |                      300000       |      56270 ns/op     |      39040 B/op   |      39 allocs/op |
| 4K                  |                    100000         |   129520 ns/op       |    59520 B/op   |     39 allocs/op |
| 8K                |                       100000       |     154120 ns/op     |     100480 B/op   |      39 allocs/op |
| 16K           |                           100000       |     238414 ns/op    |      182400 B/op  |       39 allocs/op |
| 32K           |                            30000       |     448035 ns/op    |      347200 B/op  |       59 allocs/op |
| 64K            |                           20000       |     809671 ns/op   |      676800 B/op  |       99 allocs/op |
| 128K            |                          10000       |    1547572 ns/op     |    1336000 B/op |       179 allocs/op |
| 256K               |                        5000       |    3110902 ns/op    |     2654400 B/op        |339 allocs/op |

|       Delay          |       Ops    |     Speed       |      Throughput       |    Memory     |    Allocs         | 
|-----------------|-----------|------------|-------------|---------|-------------| 
| NoReceiver      | 200000000 | 87.9 ns/op | 352.59 MB/s | 37 B/op | 0 allocs/op | 
| ConsumeDelay0   | 100000000 | 152 ns/op  | 203.78 MB/s | 20 B/op | 0 allocs/op | 
| ConsumeDelay1   | 200000000 | 90.5 ns/op | 342.47 MB/s | 37 B/op | 0 allocs/op | 
| ConsumeDelay10  | 200000000 | 86.1 ns/op | 359.98 MB/s | 37 B/op | 0 allocs/op | 
| ConsumeDelay50  | 200000000 | 85.1 ns/op | 364.14 MB/s | 37 B/op | 0 allocs/op | 
| ConsumeDelay100 | 200000000 | 82.3 ns/op | 376.85 MB/s | 37 B/op | 0 allocs/op | 
| ConsumeDelay300 | 200000000 | 81.4 ns/op | 380.61 MB/s | 37 B/op | 0 allocs/op | 
| ConsumeDelay500 | 200000000 | 78.3 ns/op | 396.16 MB/s | 37 B/op | 0 allocs/op |